# server/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DD Core API",
    "version": "0.1.0",
    "description": "Stable v1 contract for DD instance discovery, service health, account authentication, contacts, and P4 reliable text messaging/sync. P0 call endpoints remain experimental and are intentionally excluded until promoted into /api/v1.",
    "license": {
      "name": "License decision pending",
      "url": "https://dd.invalid/license-pending"
    }
  },
  "servers": [
    {
      "url": "https://{instanceHost}",
      "description": "Self-hosted DD instance",
      "variables": {
        "instanceHost": {
          "default": "dd.invalid",
          "description": "Hostname of the self-hosted instance"
        }
      }
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/v1/moments": {
      "get": {
        "operationId": "listMomentsFeed",
        "summary": "List the authenticated user's visible Moments feed",
        "tags": ["Moments"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "before", "in": "query", "required": false, "schema": {"type": "string", "format": "uuid"}},
          {"name": "authorId", "in": "query", "required": false, "description": "Restrict the feed to one author while preserving all normal visibility and privacy rules.", "schema": {"type": "string", "format": "uuid"}},
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 50, "default": 30}}
        ],
        "responses": {
          "200": {"description": "Visible Moments ordered newest first", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MomentListResponse"}}}},
          "400": {"description": "Pagination or author filter is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      },
      "post": {
        "operationId": "createMoment",
        "summary": "Publish a Moment to a server-authorized contact audience",
        "tags": ["Moments"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateMomentRequest"}}}},
        "responses": {
          "201": {"description": "Moment published", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MomentResponse"}}}},
          "400": {"description": "Text, media or visibility input is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "403": {"description": "Media ownership or audience relationship policy forbids publication", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/moments/{momentId}": {
      "get": {
        "operationId": "getMoment",
        "summary": "Get one Moment when it is currently visible to the authenticated user",
        "tags": ["Moments"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/MomentId"}],
        "responses": {
          "200": {"description": "Visible Moment", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MomentResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "404": {"description": "Moment does not exist or is not visible", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      },
      "delete": {
        "operationId": "deleteMoment",
        "summary": "Delete a Moment authored by the authenticated user",
        "tags": ["Moments"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/MomentId"}],
        "responses": {
          "200": {"description": "Moment deleted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OperationResultResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "403": {"description": "Only the author may delete this Moment", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "404": {"description": "Moment not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
    "/api/v1/moment-activity": {
      "get": {
        "operationId": "getMomentActivitySummary",
        "summary": "Get the authenticated user's durable Moments interaction inbox and unread count",
        "tags": ["Moments"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Unread count plus up to 30 recent visible likes, comments and replies", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MomentActivitySummaryResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/moment-activity/read": {
      "post": {
        "operationId": "markMomentActivityRead",
        "summary": "Mark all currently unread Moments interactions as read",
        "tags": ["Moments"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Current unread count and recent interaction inbox after marking existing activity read", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MomentActivitySummaryResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
    "/api/v1/moment-preferences": {
      "get": {
        "operationId": "listMomentRelationshipPreferences",
        "summary": "List non-default per-contact Moments privacy preferences",
        "tags": ["Moments"], "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Relationship preferences", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MomentPreferenceListResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "MomentId": {"name": "momentId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OperationResultData": {
        "type": "object",
        "additionalProperties": true
      },
      "OperationResultResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data", "requestId"],
        "properties": {"data": {"$ref": "#/components/schemas/OperationResultData"}, "requestId": {"$ref": "#/components/schemas/RequestIdValue"}}
      },
      "MomentUserPreview": {
        "type": "object", "additionalProperties": false, "required": ["id", "handle", "displayName"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "handle": {"type": "string", "pattern": "^[a-z][a-z0-9_]{2,31}$"},
          "displayName": {"type": "string", "minLength": 1, "maxLength": 80}
        }
      },
      "MomentComment": {
        "type": "object", "additionalProperties": false, "required": ["id", "author", "text", "createdAt"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "author": {"$ref": "#/components/schemas/MomentUserPreview"},
          "replyToCommentId": {"type": ["string", "null"], "format": "uuid"},
          "text": {"type": "string", "minLength": 1, "maxLength": 1000},
          "createdAt": {"type": "string", "format": "date-time"}
        }
      },
      "Moment": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "author", "text", "visibility", "mediaIds", "likeUsers", "comments", "likedByMe", "createdAt"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "author": {"$ref": "#/components/schemas/MomentUserPreview"},
          "text": {"type": "string", "maxLength": 2000},
          "visibility": {"type": "string", "enum": ["ALL_CONTACTS", "PRIVATE", "EXCLUDE"]},
          "mediaIds": {"type": "array", "maxItems": 9, "uniqueItems": true, "items": {"type": "string", "format": "uuid"}},
          "likeUsers": {"type": "array", "items": {"$ref": "#/components/schemas/MomentUserPreview"}},
          "comments": {"type": "array", "items": {"$ref": "#/components/schemas/MomentComment"}},
          "likedByMe": {"type": "boolean"},
          "createdAt": {"type": "string", "format": "date-time"}
        }
      },
      "MomentResponse": {
        "type": "object", "additionalProperties": false, "required": ["data", "requestId"],
        "properties": {"data": {"$ref": "#/components/schemas/Moment"}, "requestId": {"$ref": "#/components/schemas/RequestIdValue"}}
      },
      "MomentListData": {
        "type": "object", "additionalProperties": false, "required": ["items"],
        "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/Moment"}}}
      },
      "MomentListResponse": {
        "type": "object", "additionalProperties": false, "required": ["data", "requestId"],
        "properties": {"data": {"$ref": "#/components/schemas/MomentListData"}, "requestId": {"$ref": "#/components/schemas/RequestIdValue"}}
      },
      "MomentActivityItem": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "kind", "actor", "momentId", "createdAt", "read"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "kind": {"type": "string", "enum": ["LIKE", "COMMENT"]},
          "actor": {"$ref": "#/components/schemas/MomentUserPreview"},
          "momentId": {"type": "string", "format": "uuid"},
          "commentId": {"type": "string", "format": "uuid"},
          "commentText": {"type": "string", "maxLength": 1000},
          "createdAt": {"type": "string", "format": "date-time"},
          "read": {"type": "boolean"}
        }
      },
      "MomentActivitySummary": {
        "type": "object", "additionalProperties": false, "required": ["unreadCount", "items"],
        "properties": {
          "unreadCount": {"type": "integer", "format": "int64", "minimum": 0},
          "items": {"type": "array", "maxItems": 30, "items": {"$ref": "#/components/schemas/MomentActivityItem"}}
        }
      },
      "MomentActivitySummaryResponse": {
        "type": "object", "additionalProperties": false, "required": ["data", "requestId"],
        "properties": {"data": {"$ref": "#/components/schemas/MomentActivitySummary"}, "requestId": {"$ref": "#/components/schemas/RequestIdValue"}}
      },
...
      "CreateMomentRequest": {
        "type": "object", "additionalProperties": false,
        "properties": {
          "text": {"type": "string", "maxLength": 2000},
          "mediaIds": {"type": "array", "maxItems": 9, "uniqueItems": true, "items": {"type": "string", "format": "uuid"}},
          "visibility": {"type": "string", "enum": ["ALL_CONTACTS", "PRIVATE", "EXCLUDE"], "default": "ALL_CONTACTS"},
          "visibilityUserIds": {"type": "array", "maxItems": 500, "uniqueItems": true, "items": {"type": "string", "format": "uuid"}}
        },
        "anyOf": [
          {"required": ["text"]},
          {"required": ["mediaIds"]}
        ]
      },
...
      "MomentPreference": {
        "type": "object", "additionalProperties": false, "required": ["target", "hideTarget", "hideFromTarget", "updatedAt"],
        "properties": {
          "target": {"$ref": "#/components/schemas/MomentUserPreview"},
          "hideTarget": {"type": "boolean"},
          "hideFromTarget": {"type": "boolean"},
          "updatedAt": {"type": "string", "format": "date-time"}
        }
      },
...
      "MomentPreferenceListData": {
        "type": "object", "additionalProperties": false, "required": ["items"],
        "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/MomentPreference"}}}
      },
      "MomentPreferenceListResponse": {
        "type": "object", "additionalProperties": false, "required": ["data", "requestId"],
        "properties": {"data": {"$ref": "#/components/schemas/MomentPreferenceListData"}, "requestId": {"$ref": "#/components/schemas/RequestIdValue"}}
      },
...
      "RequestIdValue": {
        "type": "string",
        "pattern": "^req_[A-Za-z0-9]+$"
      },
      "APIError": {
        "type": "object",
        "additionalProperties": true,
        "required": ["code", "message", "requestId"],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*$"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestIdValue"
          },
          "details": {}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        }
      }
    }
  }
}